Implement basic dispatch function and exec loop.
The hard part here was creating a map of connections by database name
that's persistent for the life of the library. This HashMap can't be
held at the file level because JsValue doesn't implement the Send trait,
and so the compiler has to prove it is never shared between threads.

replicache_client.js: 13435
replicache_client.js.br: 3175
replicache_client_bg.wasm: 51263
replicache_client_bg.wasm.br: 21946