Implement Garbage Collection
We use ref counting to implement garbage collection. This is done at the
`dag::Write` layer. We do the garbage collection in `dag::Write::commit`.

- A head keeps a chunk alive.
- A ref from one alive chunk to another chunk keeps the other chunk
  alive.

When we call `set_head` we record the new and old hash so that we can
increment and decrement the ref counts later in `commit`.

When we call `put_chunk` we record the hash so that we can check for
orphaned chunks in `commit`.

Fixes #34